Update program execution command in README files and run.sh scripts
The README files and run.sh scripts have been updated to change the command for running the program from ./your_server.sh to ./your_program.sh. This change reflects the correct name of the script that
should be executed to run the program implemented in each language.

Commit message: "Update program execution command in README files and run.sh scripts"